src/requirements.txt
email-validator==2.2.0
fastapi==0.115.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib
google-auth==2.34.0
httptools
icecream
isort
openpyxl
pandas
python-dateutil
python-dotenv==1.0.1
pytz==2024.1
uvicorn[standard]==0.30.6
watchfiles
websocket
